# spec/openapi.yaml
openapi: '3.0.0'
info:
  title: Go Template API
  version: '1.0.0'
servers:
  - url: https://localhost:8080/api/v1
paths:
  /register:
    post:
      operationId:  RegisterUserHandler
      tags:
        - registration
      requestBody: 
        content: 
          application/json:
            schema: 
              type: object
              $ref: '#/components/schemas/RegisterUserRequestBody'
      responses:
        "200":
          description: "Register user"
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/CreateUserResponse' 
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Error'
        "500":
          description: Error response
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Error'

components:
  schemas:
    RegisterUserRequestBody:
      required:
        - username
        - email
        - password
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 63
        email:
          type: string
          minLength: 3
          maxLength: 253
          pattern: '^[A-Za-z0-9._+%-]+@[A-Za-z0-9.-]+[.][A-Za-z]+$'
        password:
          type: string
          minLength: 8
          maxLength: 63
          pattern: '^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]{8,}$' # Minimum eight characters, at least one letter and one number
    CreateUserResponse:
      required:
        - id
      type: object
      properties:
        id:
          type: integer
          format: int64
    Error:
      required:
        - id
        - timestamp
        - status
        - error
        - message
        - path
      type: object
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string